Initial commit of unfinished Updater
The goal is to user a scheduled workflow to execute the Updater and update the media picker

We're at the point where we can pull all the movies and episodes from Wikipedia.
Now it's just a matter of cleaning the data, storing it properly and then changing the arrays in the media picker.